Convert Ns_ThreadId() to return a uintptr_t
An uintptr_t is an integral type which is as long as a pointer, so
that it can accurately hold an opaque pthread_t whether it is
implemented as an integer type, a pointer, and on 32/64 bit
machines. This fixes many pointer/integer of different size warnings.

Log messages now print the thread ID in hex, for brevity on
64bit machines.

NB: depending on the architecture of your machine, the ABI may have
changed and require a recompile.

+++ include/nsthread.h
@@ -318,7 +318,7 @@ NS_EXTERN void Ns_ThreadExit(void *arg);
 NS_EXTERN void Ns_ThreadJoin(Ns_Thread *threadPtr, void **argPtr);
 NS_EXTERN void Ns_ThreadYield(void);
 NS_EXTERN void Ns_ThreadSetName(char *name, ...);
-NS_EXTERN int Ns_ThreadId(void);
+NS_EXTERN uintptr_t Ns_ThreadId(void);
 NS_EXTERN void Ns_ThreadSelf(Ns_Thread *threadPtr);
 NS_EXTERN char *Ns_ThreadGetName(void);
 NS_EXTERN char *Ns_ThreadGetParent(void);
